recreate folds rule
