+++ workflow/Snakefile
@@ -18,7 +18,7 @@ include: "rules/model.smk"
 rule all:
     input:
         expand(
-            rules.split_train_test.output.train,
+            rules.folds.output,
             ref=config["genome"]["build"],
             dna_type="mda",
         ),
