+++ workflow/rules/model.smk
@@ -22,7 +22,7 @@ rule get_features:
         "../scripts/get_features.py"
 
 
-rule split_train_test:
+rule folds:
     input:
         samples=expand(
             "results/get_features/{{ref}}/{donor}/{{dna_type}}/{sample}.pickle.gz",
@@ -40,11 +40,26 @@ rule split_train_test:
         min_reads=config["model"]["min_reads"],
         fold_window=config["model"]["fold_window"],
     output:
-        train="results/split_train_test/{ref}/{dna_type}/Training.csv",
-        test="results/split_train_test/{ref}/{dna_type}/Testing.csv",
+        train_features=expand(
+            "results/folds/{{ref}}/{{dna_type}}/fold_{fold}/X_train.pickle.gz",
+            fold=range(1, config["model"]["num_folds"] + 1),
+        ),
+        test_features=expand(
+            "results/folds/{{ref}}/{{dna_type}}/fold_{fold}/X_test.pickle.gz",
+            fold=range(1, config["model"]["num_folds"] + 1),
+        ),
+        train_labels=expand(
+            "results/folds/{{ref}}/{{dna_type}}/fold_{fold}/Y_train.pickle",
+            fold=range(1, config["model"]["num_folds"] + 1),
+        ),
+        test_labels=expand(
+            "results/folds/{{ref}}/{{dna_type}}/fold_{fold}/Y_test.pickle",
+            fold=range(1, config["model"]["num_folds"] + 1),
+        ),
+        label_encoder="results/folds/{ref}/{dna_type}/label_encoder.pickle",
     log:
-        "results/split_train_test/{ref}/{dna_type}.log",
+        "results/folds/{ref}/{dna_type}.log",
     conda:
         "../envs/env.yml"
     script:
-        "../scripts/split_train_test.py"
+        "../scripts/folds.py"
